# Use `aws-lc-rs`/`ring` implementation for AEAD ciphers