--- a/.all-contributorsrc
+++ b/.all-contributorsrc
@@ -571,6 +571,15 @@
       "contributions": [
         "code"
       ]
+    },
+    {
+      "login": "wyhaya",
+      "name": "wyhaya",
+      "avatar_url": "https://avatars.githubusercontent.com/u/23690145?v=4",
+      "profile": "https://github.com/wyhaya",
+      "contributions": [
+        "code"
+      ]
     }
   ],
   "contributorsPerLine": 7,