--- a/russh/src/cipher/chacha20poly1305.rs
+++ b/russh/src/cipher/chacha20poly1305.rs
@@ -36,6 +36,7 @@ impl super::Cipher for SshChacha20Poly1305Cipher {
         _: &dyn MacAlgorithm,
     ) -> Box<dyn super::OpeningKey + Send> {
         Box::new(OpeningKey(chacha20_poly1305_openssh::OpeningKey::new(
+            #[allow(clippy::unwrap_used)]
             k.first_chunk().unwrap(),
         )))
     }
@@ -49,6 +50,7 @@ impl super::Cipher for SshChacha20Poly1305Cipher {
         _: &dyn MacAlgorithm,
     ) -> Box<dyn super::SealingKey + Send> {
         Box::new(SealingKey(chacha20_poly1305_openssh::SealingKey::new(
+            #[allow(clippy::unwrap_used)]
             k.first_chunk().unwrap(),
         )))
     }
@@ -66,6 +68,7 @@ impl super::OpeningKey for OpeningKey {
     ) -> [u8; 4] {
         self.0.decrypt_packet_length(
             sequence_number,
+            #[allow(clippy::unwrap_used)]
             *encrypted_packet_length.first_chunk().unwrap(),
         )
     }
@@ -85,6 +88,7 @@ impl super::OpeningKey for OpeningKey {
             .open_in_place(
                 sequence_number,
                 ciphertext_in_plaintext_out,
+                #[allow(clippy::unwrap_used)]
                 tag.first_chunk().unwrap(),
             )
             .map_err(|_| Error::DecryptionError)
@@ -130,6 +134,7 @@ impl super::SealingKey for SealingKey {
         self.0.seal_in_place(
             sequence_number,
             plaintext_in_ciphertext_out,
+            #[allow(clippy::unwrap_used)]
             tag.first_chunk_mut().unwrap(),
         );
     }