--- a/russh/src/cipher/gcm.rs
+++ b/russh/src/cipher/gcm.rs
@@ -17,11 +17,23 @@
 
 use std::convert::TryInto;
 
-use aws_lc_rs::aead::{
-    Aad, Algorithm, BoundKey, Nonce as AeadNonce, NonceSequence, OpeningKey as AeadOpeningKey,
-    SealingKey as AeadSealingKey, UnboundKey, NONCE_LEN,
+#[cfg(not(target_arch = "wasm32"))]
+use aws_lc_rs::{
+    aead::{
+        Aad, Algorithm, BoundKey, Nonce as AeadNonce, NonceSequence, OpeningKey as AeadOpeningKey,
+        SealingKey as AeadSealingKey, UnboundKey, NONCE_LEN,
+    },
+    error::Unspecified,
 };
 use rand::RngCore;
+#[cfg(target_arch = "wasm32")]
+use ring::{
+    aead::{
+        Aad, Algorithm, BoundKey, Nonce as AeadNonce, NonceSequence, OpeningKey as AeadOpeningKey,
+        SealingKey as AeadSealingKey, UnboundKey, NONCE_LEN,
+    },
+    error::Unspecified,
+};
 
 use super::super::Error;
 use crate::mac::MacAlgorithm;
@@ -73,7 +85,7 @@ pub struct SealingKey<N: NonceSequence>(AeadSealingKey<N>);
 struct Nonce([u8; NONCE_LEN]);
 
 impl NonceSequence for Nonce {
-    fn advance(&mut self) -> Result<aws_lc_rs::aead::Nonce, aws_lc_rs::error::Unspecified> {
+    fn advance(&mut self) -> Result<AeadNonce, Unspecified> {
         let mut previous_nonce = [0u8; NONCE_LEN];
         #[allow(clippy::indexing_slicing)] // length checked
         previous_nonce.clone_from_slice(&self.0[..]);