Closes #532.

By switching the ChaCha20Poly1305, AES-GCM-128 and AES-GCM-256 ciphers from the RustCrypto implementation to `ring`'s, [my port forwarding benchmarks](https://github.com/EpicEric/sandhole/blob/main/tests/https_single_stream_download.rs) resulted in a substantial performance gain of an order of magnitude on my machine.

||With current `main` | With this PR |
|-|-|-|
|ChaCha20Poly1305|8s 18ms|662 ms|
|AES-GCM-128|6s 280ms|557ms|
|AES-GCM-256|6s 396ms|552ms|

But results seem to vary depending on your hardware, where we see much smaller or no gains. On a different machine with limited memory and less CPU power (and a different benchmark):

||With current `main` | With this PR |
|-|-|-|
|ChaCha20Poly1305|660ms|654ms|
|AES-GCM-256|803ms|615ms|

This implementation also changes the signature for the `OpeningKey::open` method to not split the tag from the cipher prematurely.

I have tested ChaCha20Poly1305 and AES-GCM-256 against the current release of russh and OpenSSH, but not AES-GCM-128, although I expect it to work the same.