@EpicEric 2025-06-28
We could switch to the aws-lc-rs implementations instead, I just wasn't able to figure out how to make it compile for Windows or WASM, so I went with ring initially.