@EpicEric 2025-06-28
Given that WebAssembly doesn't seem to be a supported platform for aws-lc-rs, I've replaced the dependency with ring for the target.

Perhaps choosing between either should be a feature flag in the same vein as rustls?